Added toolchain for alpha testing...
run the ``` join.sh ``` to cat the parts of the zip together. shasum is called -- if you don't have that, check it with what your distro provides thru the checksum file. This doesn't have that many libraries built in, so you might need to bring your own in depending on your own needs. This ideally should go in some kind of ```/opt``` folder. I give you mipsel musl gcc~!